pyproject.toml
[project]
name = "ffwrapped-be"
version = "0.1.0"
description = "Backend for fantasy football wrapped project"
authors = [
    {name = "kevniu95",email = "kniu95@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
packages = [
    {include = "ffwrapped_be"}
]
dependencies = [
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "ratelimit (>=2.2.1,<3.0.0)",
    "mechanize (>=0.4.10,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
